Address the comment
